feat(notif): Uses signals for topics.
Moves TopicFollowed in the notification module and
creates all signals/receivers used in the existing
code.
