+++ zds/forum/admin.py
@@ -2,8 +2,7 @@
 
 from django.contrib import admin
 
-from .models import Category, Forum, Topic, Post, TopicRead, TopicFollowed
-
+from zds.forum.models import Category, Forum, Topic, Post, TopicRead
 
 admin.site.register(Category)
 admin.site.register(Forum)
@@ -11,4 +10,3 @@ admin.site.register(Forum)
 admin.site.register(Topic)
 admin.site.register(Post)
 admin.site.register(TopicRead)
-admin.site.register(TopicFollowed)
