+++ zds/notification/admin.py
@@ -1,0 +1,7 @@
+# coding: utf-8
+
+from django.contrib import admin
+
+from .models import TopicFollowed
+
+admin.site.register(TopicFollowed)
