+++ zds/settings.py
@@ -186,6 +186,7 @@ INSTALLED_APPS = (
     'zds.member',
     'zds.featured',
     'zds.search',
+    'zds.notification',
     # Uncomment the next line to enable the admin:
     'django.contrib.admin',
     # Uncomment the next line to enable admin documentation:
