+++ zds/utils/templatetags/interventions.py
@@ -1,16 +1,18 @@
 # coding: utf-8
 
-from datetime import datetime, timedelta
 import time
+from datetime import datetime, timedelta
 
 from django import template
 from django.db.models import F
 from django.utils.translation import ugettext_lazy as _
 
-from zds.forum.models import TopicFollowed, never_read as never_read_topic, Post, TopicRead
+from zds.forum.models import Post
 from zds.mp.models import PrivateTopic
-from zds.utils.models import Alert
+from zds.notification.models import TopicRead, TopicFollowed
+from zds.notification.models import never_read as never_read_topic
 from zds.tutorialv2.models.models_database import ContentRead, ContentReaction
+from zds.utils.models import Alert
 
 register = template.Library()
